fix for macOS

+++ tpl.sys/backup/backup.sh
@@ -2,21 +2,51 @@
 umask 022
 LANG=C
 export LANG
-ext='=:='
 
-tmp=/tmp/`basename $0`.$$
+PATH=/usr/bin:/bin; export PATH
+for v in $(printenv | while read vv; do echo ${vv%%=*}; done); do
+    case "$v" in
+        (HOME|PATH|PWD|TERM|TMPDIR|TZ|USER) ;;
+        (*) unset $v ;;
+    esac
+done
+
+gfind=
+for gf in $(
+    for x in gfind find; do
+        which -a $x
+        for p in /usr/local/bin $HOME/sys/local/bin /opt/homebrew/bin; do
+            [ ! -x $p/$x ] || echo $p/$x
+        done
+    done); do
+    if $gf --version 2>&1 | grep 'GNU find' > /dev/null 2> /dev/null; then
+        gfind=$gf
+        break
+    fi
+done
+if [ -z "$gfind" ]; then
+    echo "ERROR: can not find GNU find"
+    exit 1
+fi
+
+etc=$(cd /etc && pwd)
+tmp=${TMPDIR:-/tmp}/$(basename $0).$$
 trap "rm -rf $tmp.*; exit 1" 1 2 3 15
 rm -rf $tmp.*
+ext='=:='
 
 ###
 ### shell functions
 ###
+echon() {
+    printf '%s' "$*"
+}
 
 read_ans() {
     msg="$1"
 
     while :; do
-        echo -n "$msg"
+        echon "$msg"
         if [ -z "$answer" ]; then
             read ans < /dev/tty
         else
@@ -48,36 +78,36 @@ epoch_touch() {
 
 backup_file() {
     etcorig="$1"
-    etcfile=`dirname "$etcorig"`/`basename "$etcorig" .orig`
-    bang=`echo "$etcfile" | sed -e 's#/#!#g'`
+    etcfile=$(dirname "$etcorig")/$(basename "$etcorig" .orig)
+    bang=$(echo "$etcfile" | sed -e 's#/#!#g')
     etcorigrm=0
 
     if [ -e "$etcorig" -a '(' ! -f "$etcorig" -a ! -L "$etcorig" ')' ]; then
         echo "WARNING: $etcorig is not a regular file, skipping"
-        echo -n "WARNING: "; ls -ld "$etcorig"
+        echon "WARNING: "; ls -ld "$etcorig"
         echo
         return
     fi
 
     if [ -e "$etcfile" -a '(' ! -f "$etcfile" -a ! -L "$etcfile" ')' ]; then
         echo "WARNING: $etcfile is not a regular file, skipping"
-        echo -n "WARNING: "; ls -ld "$etcfile"
+        echon "WARNING: "; ls -ld "$etcfile"
         echo
         continue
     fi
 
     if [ -e "orig/$bang" -a -e "orig/${bang}${ext}" ]; then
         echo "ERROR: duplicate backup file: $etcfile"
-        echo -n "ERROR: "; ls -ld "orig/$bang"
-        echo -n "ERROR: "; ls -ld "orig/${bang}${ext}"
+        echon "ERROR: "; ls -ld "orig/$bang"
+        echon "ERROR: "; ls -ld "orig/${bang}${ext}"
         rm -rf $tmp.*
         exit 1
     fi
 
     if [ -e "new/$bang" -a -e "new/${bang}${ext}" ]; then
         echo "ERROR: duplicate backup file: $etcfile"
-        echo -n "ERROR: "; ls -ld "new/$bang"
-        echo -n "ERROR: "; ls -ld "new/${bang}${ext}"
+        echon "ERROR: "; ls -ld "new/$bang"
+        echon "ERROR: "; ls -ld "new/${bang}${ext}"
         rm -rf $tmp.*
         exit 1
     fi
@@ -86,10 +116,10 @@ backup_file() {
 
         if [ -e "orig/$bang" ]; then
             echo "NOTICE: backup file exists: $etcfile"
-            echo -n "NOTICE: "; ls -ld "orig/$bang"
+            echon "NOTICE: "; ls -ld "orig/$bang"
             etcorigrm=1
         else
-            if [ `find "$etcorig" -printf "%m"` -eq 0 ]; then
+            if [ $($gfind "$etcorig" -printf "%m") -eq 0 ]; then
                 cp -pi "$etcfile" "orig/$bang"
                 : > "orig/$bang"
                 epoch_touch "orig/$bang"
@@ -97,7 +127,7 @@ backup_file() {
                 cp -pi "$etcorig" "orig/$bang"
             fi
             echo "INFO: backup file created: $etcorig"
-            echo -n "INFO: "; ls -ld "orig/$bang"
+            echon "INFO: "; ls -ld "orig/$bang"
             echo
             etcorigrm=1
         fi
@@ -105,7 +135,7 @@ backup_file() {
         if [ ! -e "new/$bang" ]; then
             cp -pi "$etcfile" "new/$bang"
             echo "INFO: backup file created: $etcfile"
-            echo -n "INFO: "; ls -ld "new/$bang"
+            echon "INFO: "; ls -ld "new/$bang"
             echo
         fi
 
@@ -113,45 +143,45 @@ backup_file() {
 
         if [ -e "orig/${bang}${ext}" ]; then
             echo "NOTICE: backup file exists: $etcfile"
-            echo -n "NOTICE: "; ls -ld "orig/${bang}${ext}"
+            echon "NOTICE: "; ls -ld "orig/${bang}${ext}"
             etcorigrm=1
         else
-            if [ `find "$etcorig" -printf "%m"` -eq 0 ]; then
+            if [ $($gfind "$etcorig" -printf "%m") -eq 0 ]; then
                 : > $tmp.epoch
                 epoch_touch $tmp.epoch
-                find "$etcfile" -printf "%m %u:%g " > "orig/${bang}${ext}"
-                find $tmp.epoch -printf \
+                $gfind "$etcfile" -printf "%m %u:%g " > "orig/${bang}${ext}"
+                $gfind $tmp.epoch -printf \
                 "%s %TY-%Tm-%Td %TT $etcfile\n" >> "orig/${bang}${ext}"
                 touch -h -r $tmp.epoch "orig/${bang}${ext}"
             else
-                find "$etcorig" -printf \
+                $gfind "$etcorig" -printf \
                 "%m %u:%g %s %TY-%Tm-%Td %TT $etcfile" > "orig/${bang}${ext}"
                 if [ -h "$etcorig" ]; then
-                    find "$etcorig" -printf " -> %l\n" >> "orig/${bang}${ext}"
+                    $gfind "$etcorig" -printf " -> %l\n" >> "orig/${bang}${ext}"
                 else
                     echo >> "orig/${bang}${ext}"
                 fi
                 touch -h -r "$etcorig" "orig/${bang}${ext}"
             fi
             echo "INFO: backup file created: $etcorig"
-            echo -n "INFO: "; ls -ld "orig/${bang}${ext}"
-            echo -n "INFO: "; cat "orig/${bang}${ext}"
+            echon "INFO: "; ls -ld "orig/${bang}${ext}"
+            echon "INFO: "; cat "orig/${bang}${ext}"
             echo
             etcorigrm=1
         fi
 
         if [ ! -e "new/${bang}${ext}" ]; then
-            find "$etcfile" -printf \
+            $gfind "$etcfile" -printf \
             "%m %u:%g %s %TY-%Tm-%Td %TT $etcfile" > "new/${bang}${ext}"
             if [ -h "$etcfile" ]; then
-                find "$etcfile" -printf " -> %l\n" >> "new/${bang}${ext}"
+                $gfind "$etcfile" -printf " -> %l\n" >> "new/${bang}${ext}"
             else
                 echo >> "new/${bang}${ext}"
             fi
             touch -h -r "$etcfile" "new/${bang}${ext}"
             echo "INFO: backup file created: $etcfile"
-            echo -n "INFO: "; ls -ld "new/${bang}${ext}"
-            echo -n "INFO: "; cat "new/${bang}${ext}"
+            echon "INFO: "; ls -ld "new/${bang}${ext}"
+            echon "INFO: "; cat "new/${bang}${ext}"
             echo
         fi
 
@@ -166,7 +196,7 @@ backup_dir_check() {
     a=$1
     b=$2
 
-    find $a -type f -printf "%P\n" | sort -u | while read bang; do
+    $gfind $a -type f -printf "%P\n" | sort -u | while read bang; do
         if [ ! -f "$b/$bang" ]; then
             echo "WARNING: missing backup file: $b/$bang"
         fi
@@ -174,14 +204,14 @@ backup_dir_check() {
 }
 
 backup_diff_check() {
-    find new -type f -printf "%P\n" | \
+    $gfind new -type f -printf "%P\n" | \
         sed -e 's/'"${ext}"'$//' | sort -u | while read bang; do
-        etcfile=`echo "$bang" | sed -e 's#!#/#g'`
+        etcfile=$(echo "$bang" | sed -e 's#!#/#g')
 
         if [ -e "new/$bang" -a -e "new/${bang}${ext}" ]; then
             echo "ERROR: duplicate backup file: $etcfile"
-            echo -n "ERROR: "; ls -ld "new/$bang"
-            echo -n "ERROR: "; ls -ld "new/${bang}${ext}"
+            echon "ERROR: "; ls -ld "new/$bang"
+            echon "ERROR: "; ls -ld "new/${bang}${ext}"
             rm -rf $tmp.*
             exit 1
         fi
@@ -199,16 +229,16 @@ backup_diff_check() {
                 cp -p "$etcfile" "new/$bang"
                 if [ $? -eq 0 ]; then
                     echo "INFO: backup file updated: $etcfile"
-                    echo -n "INFO: "; ls -ld "new/$bang"
+                    echon "INFO: "; ls -ld "new/$bang"
                     echo
                 fi
             fi
 
         else
 
-            find "$etcfile" -printf "%m %u:%g %s %TY-%Tm-%Td %TT %p" > $tmp.etc
+            $gfind "$etcfile" -printf "%m %u:%g %s %TY-%Tm-%Td %TT %p" > $tmp.etc
             if [ -h "$etcfile" ]; then
-                find "$etcfile" -printf " -> %l\n" >> $tmp.etc
+                $gfind "$etcfile" -printf " -> %l\n" >> $tmp.etc
             else
                 echo >> $tmp.etc
             fi
@@ -220,18 +250,18 @@ backup_diff_check() {
             read_ans "NOTICE: Do you want to update new/$bang [y/N]? "
 
             if [ $? -eq 0 ]; then
-                find "$etcfile" -printf \
+                $gfind "$etcfile" -printf \
                 "%m %u:%g %s %TY-%Tm-%Td %TT %p" > "new/${bang}${ext}"
                 if [ -h "$etcfile" ]; then
-                    find "$etcfile" -printf " -> %l\n" >> "new/${bang}${ext}"
+                    $gfind "$etcfile" -printf " -> %l\n" >> "new/${bang}${ext}"
                 else
                     echo >> "new/${bang}${ext}"
                 fi
                 touch -h -r "$etcfile" "new/${bang}${ext}"
                 if [ $? -eq 0 ]; then
                     echo "INFO: backup file updated: $etcfile"
-                    echo -n "INFO: "; ls -ld "new/${bang}${ext}"
-                    echo -n "INFO: "; cat "new/${bang}${ext}"
+                    echon "INFO: "; ls -ld "new/${bang}${ext}"
+                    echon "INFO: "; cat "new/${bang}${ext}"
                     echo
                 fi
             fi
@@ -244,9 +274,9 @@ backup_diff_check() {
 ### main
 ###
 sudo true || exit 1
-find /etc -xdev -name '*.orig' 1> $tmp.out 2> $tmp.err
-egrep    '^find: .*: Permission denied$' $tmp.err > $tmp.err1
-egrep -v '^find: .*: Permission denied$' $tmp.err > $tmp.err2
+$gfind $etc -xdev -name '*.orig' 1> $tmp.out 2> $tmp.err
+grep -E    '(^|/)g?find: .*: Permission denied$' $tmp.err > $tmp.err1
+grep -E -v '(^|/)g?find: .*: Permission denied$' $tmp.err > $tmp.err2
 
 if [ -s $tmp.err2 ]; then
     echo "ERROR: unexpected error as follows:"
@@ -255,9 +285,9 @@ if [ -s $tmp.err2 ]; then
     exit 1
 fi
 
-sed -n -e 's/^find: `\(.*\)'"'"': Permission denied$/\1/p' $tmp.err1 \
+sed -n -E -e 's/(^|\/)g?find: `(.*)'"'"': Permission denied$/\2/p' $tmp.err1 \
 | while read x; do
-    sudo find "$x" -xdev -name '*.orig' -print >> $tmp.err3
+    sudo $gfind "$x" -xdev -name '*.orig' -print >> $tmp.err3
 done
 if [ -s $tmp.err3 ]; then
     echo "ERROR: can not access following files:"
@@ -271,7 +301,7 @@ sort -u < $tmp.out | while read x; do
     backup_file "$x"
 done
 
-sudo find /etc -xdev -name '*.orig' 1> $tmp.out 2> /dev/null
+sudo $gfind $etc -xdev -name '*.orig' 1> $tmp.out 2> /dev/null
 if [ -s $tmp.out ]; then
     echo 'WARNING: *.orig files not yet removed:'
     cat $tmp.out
